+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.9] - 2026-01-20
+
+### Fixed
+- Fixed creation of folders under root
+- Fixed folder move operations to properly support moving to root level
+- Improved path normalization for folder operations
+
 ## [0.2.8] - 2026-01-19
 
 ### Fixed